services/backend/requirements.txt
google-adk
google-cloud-aiplatform
google-cloud-firestore
googleapis-common-protos
protobuf
vertexai
PyYAML
a2a
aiohttp
